feat: move tax util functions and schema to generic CommonOrder
We expect other plugins to need the same transformations
