+++ imports/plugins/core/taxes/server/no-meteor/mutations/getFulfillmentGroupTaxes.js
@@ -1,7 +1,8 @@
 import Logger from "@reactioncommerce/logger";
 import ReactionError from "@reactioncommerce/reaction-error";
+import { CommonOrder } from "/imports/plugins/core/orders/server/no-meteor/simpleSchemas";
 import { getActiveTaxServiceForShop } from "../registration";
-import { TaxServiceOrderInput, TaxServiceResult } from "../../../lib/simpleSchemas";
+import { TaxServiceResult } from "../../../lib/simpleSchemas";
 
 /**
  * @summary Returns all taxes that apply to a provided order, delegating to a more specific
@@ -17,7 +18,7 @@ import { TaxServiceOrderInput, TaxServiceResult } from "../../../lib/simpleSchem
  */
 export default async function getFulfillmentGroupTaxes(context, { order, forceZeroes }) {
   try {
-    TaxServiceOrderInput.validate(order);
+    CommonOrder.validate(order);
   } catch (error) {
     Logger.error("Invalid order input provided to getFulfillmentGroupTaxes", error);
     throw new ReactionError("internal-error", "Error while calculating taxes");
