+++ imports/plugins/core/taxes/server/no-meteor/startup.js
@@ -1,61 +1,5 @@
 import { isEqual } from "lodash";
-
-/**
- * @param {Object} cart A cart
- * @param {Object} group The cart fulfillment group
- * @param {Object} context App context
- * @returns {Object} Valid TaxServiceOrderInput from a cart group
- */
-async function buildOrderInputForTaxCalculation(cart, group, context) {
-  const { collections } = context;
-  const { currencyCode } = cart;
-
-  let items = group.itemIds.map((itemId) => cart.items.find((item) => item._id === itemId));
-  items = items.filter((item) => !!item); // remove nulls
-
-  // We also need to add `subtotal` on each item, based on the current price of that item in
-  // the catalog. `getFulfillmentGroupTaxes` uses subtotal prop to calculate the tax.
-  items = items.map((item) => ({
-    _id: item._id,
-    isTaxable: item.isTaxable,
-    parcel: item.parcel,
-    price: item.price,
-    quantity: item.quantity,
-    shopId: item.shopId,
-    subtotal: {
-      amount: item.price.amount * item.quantity,
-      currencyCode
-    },
-    taxCode: item.taxCode,
-    variantId: item.variantId
-  }));
-
-  const { address, shipmentMethod, shopId, type: fulfillmentType } = group;
-  const shop = await collections.Shops.findOne({ _id: shopId });
-
-  return {
-    currencyCode: cart.currencyCode,
-    fulfillmentPrices: {
-      handling: {
-        amount: (shipmentMethod && shipmentMethod.handling) || 0,
-        currencyCode
-      },
-      shipping: {
-        amount: (shipmentMethod && shipmentMethod.rate) || 0,
-        currencyCode
-      },
-      total: {
-        amount: shipmentMethod ? ((shipmentMethod.handling || 0) + (shipmentMethod.rate || 0)) : 0,
-        currencyCode
-      }
-    },
-    fulfillmentType,
-    items,
-    originAddress: (shop && Array.isArray(shop.addressBook) && shop.addressBook[0]) || null,
-    shippingAddress: address || null,
-    shopId
-  };
-}
+import xformCartGroupToCommonOrder from "/imports/plugins/core/cart/server/no-meteor/util/xformCartGroupToCommonOrder";
 
 /**
  * @summary Returns `cart.items` with tax-related props updated on them
@@ -65,7 +9,7 @@ async function buildOrderInputForTaxCalculation(cart, group, context) {
  */
 async function getUpdatedCartItems(cart, context) {
   const taxResultsByGroup = await Promise.all(cart.shipping.map(async (group) => {
-    const order = await buildOrderInputForTaxCalculation(cart, group, context);
+    const order = await xformCartGroupToCommonOrder(cart, group, context);
     return context.mutations.getFulfillmentGroupTaxes(context, { order, forceZeroes: false });
   }));
